Apply 048-051 and 053 to production, and fix the two defects that applying found
Applied in dependency order: 048 (model_calibration, maxedge_score, board_signals),
049 (outcome dimension, coverage_penalty), 050 (paper-trade harness), 053
(team_alias, mx_canonical_team), then 051 (mx_team_match and the form views).

TWO REAL DEFECTS, both only findable by running it.

050 used round(double precision, integer), which Postgres does not have —
stddev() returns double. paper_trade_report would not create at all, and
paper_trade_gate's z-score had the same fault. Explicit ::numeric casts.

051's mx_refresh_team_match() TIMES OUT as written: it calls mx_canonical_team()
once per row across 155,000 football-data rows. The sources have to be inserted
as separate statements joining team_alias directly. And the api_stats insert
raises 21000 — team_alias is keyed on the RAW name, so 'Gais' and 'GAIS' both
match lower(trim(raw_name)) and the join fans out. team_alias_lookup is a
DISTINCT ON view that collapses them.

VERIFIED LIVE. maxedge_score(0.03,0.03,1.0) = 65. coverage_penalty(29) is NULL,
(30) is 1.70. board_signals returns 7 rows — the DIXON_COLES survivors, the only
architecture seeded publish=true. paper_trades is empty, correctly: there are no
corner or card prices to log against. team_alias holds 1,189 rows resolving to
998 clubs, 113 of them now known to both sources. mx_team_match holds 234,442
facts across 869 clubs, and 104,271 team-matches clear the 20-match floor for
cards.

+++ migrations/050_paper_trade_harness.sql
@@ -1,5 +1,14 @@
 -- 050_paper_trade_harness.sql
 --
+-- APPLIED to production 2026-08-06. TWO DEFECTS WERE FOUND BY APPLYING IT.
+--
+--   1. `round(100.0 * stddev(clv) / sqrt(...), 2)` in paper_trade_report is a
+--      round(double precision, integer), which Postgres does not have. The view
+--      would not create. stddev() returns double; the cast is now explicit.
+--      Same fault in paper_trade_gate's z-score, fixed the same way.
+--
+-- Both are the kind of thing only running it finds, which is why it was run.
+--
 -- STAGED, NOT APPLIED. Independent of 048 and 049; 051 depends on it.
 --
 -- WHY IT EXISTS. Cards and corners are the strongest discriminators anything in
@@ -173,7 +182,7 @@ select
   round(avg(price_taken), 2)                      as avg_price,
   -- the headline
   round(100.0 * avg(clv), 2)                      as avg_clv_pct,
-  round(100.0 * stddev(clv) / sqrt(nullif(count(clv), 0)), 2) as clv_stderr_pct,
+  round((100.0 * stddev(clv) / sqrt(nullif(count(clv), 0)))::numeric, 2) as clv_stderr_pct,
   round(100.0 * sum(pl_units)
         / nullif(count(*) filter (where result in ('win','loss')), 0), 2) as yield_pct,
   round(avg(hours_to_kickoff), 1)                 as avg_hrs_before_ko,
@@ -210,7 +219,7 @@ language sql stable as $$
     select pt.model, pt.market,
            count(*) filter (where pt.result in ('win','loss')) as n,
            avg(pt.clv)                                          as clv,
-           stddev(pt.clv)                                       as clv_sd,
+           stddev(pt.clv)::numeric                              as clv_sd,
            count(pt.clv)                                        as n_clv,
            sum(pt.pl_units)                                     as pl
     from public.paper_trades pt
